Attempted Vue Slideshow
Yes I know generic WebApplication name is generic

Also, this is "how to trick your .DotNET application into running Javascript"
Will hopefully be transforming the experience into a talk one day, but will have to go back through it to refresh memory after getting the basics of the Vue ecosystem down.

So, the script works, as intended, in theory.

It is changing the value of imgs (called that as to not conflict with the img html tag). However, the script runs indefinitly, slowing the page down and preventing the image from updating reactive to the changes.

Must figure out a way to get a slideshow to run indefinitely, without screwing up the user experience, and making the page otherwise unusable. But, here is some code, and some images I made.