More Comments, More Rewriting, More Questions
Trying to persist state through the application, and figure out how to do a basic data store, before diving into vuex.

Will likely need to rewrite the app to use Webpack because of the availibility of information for it verses using .NET Core. Not sure yet if I want to do that or figure this one out