Cleared Vue Project
I know it looks drastic, but since this is just practice I can do it.

Turns out I wasn't using the best organizational practice for Vue projects and feel that I will have better luck with it if I restructure. Thankfully, I hadn't gotten really far outside of making a successful implementation of a slideshow.

Files are in another folder for now, if I return to the original project