Reorganized, Hacked around and brought back a little
I tried removing Todos from the project and putting it in the template in App.vue and forcibly invoking created on page load. Caused infinite loop of requests, but did show results from JSON Placeholder. AddTodos works as expected

Brought Todos back and moved the emit into a method to make it "reactive". Not sure if it worked or not.

Delete Todos does not work. Not sure if the emit is working as intended.

Note: original source code is copied DIRECTLY FROM A TUTORIAL!
I double and triple checked that each file was exactly the same, before making my hacky changes.
However, my project did NOT work as it was supposed to!

I am trying to fix this, so that afterwards I can build my own project, secure in the knowledge that I know a little more what I am doing. This is meant to build confidence and increase my understanding about project architecture. Not sure if it is helping or not