Thought Process
Committing this mess of comments so you can see my thought process, so you know I am not making these changes on a whim, or reverting to a previous state to follow the tutorial. I am genuinely trying to understand how Vue is processing the data in this tutorial (specifically the todos in Todos.vue, which seem to be the lynchpin issue), where it pulls from and how best to go about debugging it.