Added Router
Had to rewrite App.vue into Home.vue.

Still dont know what is going on with my requests to JSONPlaceholder.

created() does not run, at all, even though it is supposed to be linked to the created() lifecycle event in vue.js

addTodos works as expected, however, due to JSONPlaceholder, all added todos, have 201 as their id. This is how JSONPlaceholder is written and not under my control

deleteTodos(id) runs, however, always returns a 404. The id that it sends is either 1: not linked to the event; though I have verified that the emits for del-todo return the id of 201. Or 2: it overrides the id with the value of undefined. Or 3: I need to use async/await with all of my requests due to the nature of http requests (plus the sync reqs are slow...). or 4: there is a problem with using the https vs http due to not having a https cert on my localhost.