Thought I was getting closer
Reduced my state to Pictures array with dynamically mapped picture objects with properties.
Cant figure out how to properly map to a getter for OpenPictures to work due to need to filter by picture_id
Probably need to learn how to effectively use Props in Vue to pass references around components

So, got it to compile, and broke it again. Still not sure if I know what I am doing